+++ src/vector_macro.rs
@@ -1,4 +1,8 @@
-/// R like concatenate
+use std::convert;
+
+pub type Vector = Vec<f64>;
+
+/// R like concatenate (Type: Vec\<f64\>)
 ///
 /// # Examples
 /// ```
@@ -36,7 +40,7 @@ macro_rules! c {
     }
 }
 
-/// R like seq macro (Not precise)
+/// R like seq macro
 ///
 /// # Examples
 /// ```
@@ -44,6 +48,7 @@ macro_rules! c {
 /// use peroxide::*;
 ///
 /// assert_eq!(seq!(1,10,1), c!(1,2,3,4,5,6,7,8,9,10));
+/// assert_eq!(seq!(1,10,1), seq!(1;10;1));
 /// ```
 #[macro_export]
 macro_rules! seq {
@@ -65,4 +70,66 @@ macro_rules! seq {
             v
         }
     };
+    ( $start:expr; $end:expr; $step:expr ) => {
+        seq!($start, $end, $step)
+    }
+}
+
+/// zeros - like numpy
+///
+/// # Examples
+/// ```
+/// extern crate peroxide;
+/// use peroxide::*;
+///
+/// let a = zeros!(4);
+/// assert_eq!(a, c!(0,0,0,0));
+/// ```
+#[macro_export]
+macro_rules! zeros {
+    ( $n:expr ) => {
+        vec![0f64; $n]
+    };
+}
+
+pub trait FPVector {
+    fn fmap<F>(&self, f: F) -> Vector where F: Fn(f64) -> f64;
+    fn reduce<F, T>(&self, init: T, f: F) -> f64
+        where F: Fn(f64, f64) -> f64,
+              T: convert::Into<f64>;
+}
+
+impl FPVector for Vector {
+    /// fmap for Vector
+    ///
+    /// # Examples
+    /// ```
+    /// extern crate peroxide;
+    /// use peroxide::*;
+    ///
+    /// let a = c!(1,2,3,4,5);
+    /// assert_eq!(a.fmap(|x| x*2f64), seq!(2,10,2));
+    /// ```
+    fn fmap<F>(&self, f: F) -> Vector where F: Fn(f64) -> f64 {
+        self.clone().into_iter().map(|x| f(x)).collect::<Vector>()
+    }
+
+    /// reduce for Vector
+    ///
+    /// # Examples
+    /// ```
+    /// extern crate peroxide;
+    /// use peroxide::*;
+    ///
+    /// let a = seq!(1,100,1);
+    /// assert_eq!(a.reduce(0, |x,y| x + y), 5050f64);
+    /// ```
+    fn reduce<F, T>(&self, init: T, f: F) -> f64
+        where F: Fn(f64, f64) -> f64,
+              T: convert::Into<f64> {
+        self.clone().into_iter().fold(
+            init.into(),
+            |x,y| f(x,y),
+        )
+    }
 }